Relocated the `gt_set_iteration_order()` function.
Before the function was in the simplify module, but it made more sense to out it in the `map_order` module, where the transformation is implemented.
