+++ src/gt4py/next/program_processors/runners/dace_fieldview/transformations/__init__.py
@@ -22,7 +22,7 @@ from .gpu_utils import (
 from .loop_blocking import LoopBlocking
 from .map_fusion_parallel import MapFusionParallel
 from .map_fusion_serial import MapFusionSerial
-from .map_orderer import MapIterationOrder
+from .map_orderer import MapIterationOrder, gt_set_iteration_order
 from .map_promoter import SerialMapPromoter
 from .simplify import (
     GT_SIMPLIFY_DEFAULT_SKIP_SET,
@@ -30,7 +30,6 @@ from .simplify import (
     GT4PyMoveTaskletIntoMap,
     gt_inline_nested_sdfg,
     gt_reduce_distributed_buffering,
-    gt_set_iteration_order,
     gt_simplify,
     gt_substitute_compiletime_symbols,
 )
