+++ src/gt4py/next/program_processors/runners/dace_fieldview/transformations/map_orderer.py
@@ -16,6 +16,36 @@ from gt4py.next import common as gtx_common
 from gt4py.next.program_processors.runners.dace_fieldview import utility as gtx_dace_fieldview_util
 
 
+def gt_set_iteration_order(
+    sdfg: dace.SDFG,
+    leading_dim: Optional[
+        Union[str, gtx_common.Dimension, list[Union[str, gtx_common.Dimension]]]
+    ] = None,
+    validate: bool = True,
+    validate_all: bool = False,
+) -> Any:
+    """Set the iteration order of the Maps correctly.
+
+    Modifies the order of the Map parameters such that `leading_dim`
+    is the fastest varying one, the order of the other dimensions in
+    a Map is unspecific. `leading_dim` should be the dimensions were
+    the stride is one.
+
+    Args:
+        sdfg: The SDFG to process.
+        leading_dim: The leading dimensions.
+        validate: Perform validation at the end of the function.
+        validate_all: Perform validation also on intermediate steps.
+    """
+    return sdfg.apply_transformations_once_everywhere(
+        MapIterationOrder(
+            leading_dims=leading_dim,
+        ),
+        validate=validate,
+        validate_all=validate_all,
+    )
+
+
 @dace_properties.make_properties
 class MapIterationOrder(dace_transformation.SingleStateTransformation):
     """Modify the order of the iteration variables.
